requirements.txt
fastapi==0.88.0
redis-om==0.1.1
uvicorn==0.20.0
python-decouple==3.6